chore: rename file and vars
