+++ src/errors/SgidError.ts
